pyproject.toml
[project]
name = "ai-telegram"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.5",
    "langchain-openai>=1.0.2",
    "langgraph>=1.0.3",
    "openai>=2.7.2",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
    "requests>=2.32.5",
]
